feat : features/auth 1차 작업 완료
- shakeAnimation [shared/style 이동]
