+++ client/pages/join/index.tsx
@@ -1,27 +1,1 @@
-import styled from 'styled-components';
-
-import SignUp from '@/components/Auth/SignUp';
-
-import { styleMixin } from '@/src/shared/styles';
-
-function SignupPage() {
-  return (
-    <JoinStyle>
-      <SignUp />
-    </JoinStyle>
-  );
-}
-
-export default SignupPage;
-
-const JoinStyle = styled.div`
-  z-index: 1;
-  position: absolute;
-  top: 0;
-  left: 0;
-  ${styleMixin.Flex()};
-  height: 100vh;
-  width: 100vw;
-  background-color: #31404e;
-  color: ${({ theme }) => theme.colors.text};
-`;
+export { default } from '@/src/pages/join';
