+++ client/src/features/auth/Login.tsx
@@ -2,9 +2,10 @@ import { useState, useEffect } from 'react';
 
 import { useRouter } from 'next/router';
 
-import styled, { css, keyframes } from 'styled-components';
+import styled, { css } from 'styled-components';
 
 import { InputField } from '@/src/shared/components/input';
+import { shakeAnimation } from '@/src/shared/styles';
 
 import useLoginLogic from './hooks/useLoginLogic';
 
@@ -67,41 +68,6 @@ function Login() {
 export default Login;
 
 // 오류가 발생했을시 흔드는 효과
-const shakeAnimation = keyframes`
-  0% {
-    transform: translateX(0);
-  }
-  10% {
-    transform: translateX(-10px);
-  }
-  20% {
-    transform: translateX(10px);
-  }
-  30% {
-    transform: translateX(-10px);
-  }
-  40% {
-    transform: translateX(10px);
-  }
-  50% {
-    transform: translateX(-10px);
-  }
-  60% {
-    transform: translateX(10px);
-  }
-  70% {
-    transform: translateX(-10px);
-  }
-  80% {
-    transform: translateX(10px);
-  }
-  90% {
-    transform: translateX(-10px);
-  }
-  100% {
-    transform: translateX(0);
-  }
-`;
 
 const Form = styled.form<{ $shakeTrigger: boolean }>`
   display: flex;
