+++ client/src/shared/styles/index.ts
@@ -1,5 +1,5 @@
 export { default as V } from './variables';
 export {
-  fadeIn, fadeOut, slideIn, progressBar,
+  fadeIn, fadeOut, slideIn, progressBar, shakeAnimation,
 } from './animations';
 export * as styleMixin from './mixin';
